requirements.txt
supabase
python-dotenv
pytest
